Convert to single quote strings, move K8sResource
